+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # dprint-intellij-plugin Changelog
 
 ## [Unreleased]
+- Handle execution exceptions when running can format
 
 ## [0.3.2]
 - Fix intermittent lock up when running format